implemente las exception creadas y cree una nueva

+++ src/main/java/com/gimnasio/demo/Service/UsuarioServicio.java
@@ -2,6 +2,7 @@ package com.gimnasio.demo.Service;
 
 import com.gimnasio.demo.Controller.UsuarioController;
 import com.gimnasio.demo.DTO.UsuarioRegistroDTO;
+import com.gimnasio.demo.Exceptions.UsuarioNoEncontradoException;
 import com.gimnasio.demo.Model.Tarjeta;
 import com.gimnasio.demo.Repository.TarjetaRepositorio;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,8 +25,15 @@ public class UsuarioServicio {
     @Autowired
     private TarjetaRepositorio tarjetaRepositorio;
 
-    public Optional<Usuario> buscarUsuarioPorID(Long id){
-        return usuarioRepositorio.findById(id);
+    public Optional<Usuario> buscarUsuarioPorID(Long id) throws UsuarioNoEncontradoException{
+        Optional<Usuario> usuario;
+        if(usuarioRepositorio.existsById(id)){
+             usuario=usuarioRepositorio.findById(id);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
+        }
+
+        return usuario;
     }
 
     public Usuario convertidorDTO(UsuarioRegistroDTO usu){
@@ -46,20 +54,24 @@ public class UsuarioServicio {
         return b;
     }
 
-    public void deleteUser(Usuario usuario){usuarioRepositorio.delete(usuario);}
+    public void eliminarUsuarioPorID(Long id) throws UsuarioNoEncontradoException{
 
-    /// IMPLEMENTAR EXCEPTION/////////////////////////////////////////////////////////////////////////////////////
-    public void eliminarUsuarioPorID(Long id){
-        usuarioRepositorio.deleteById(id);
+        if(usuarioRepositorio.existsById(id)){
+            usuarioRepositorio.deleteById(id);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
+        }
     }
 
     public List<Usuario> listarUsuarios(){
         return usuarioRepositorio.findAll();
     }
 
-    public void editarUsuario(Long id, Usuario usuario){
+    public void editarUsuario(Long id, Usuario usuario) throws UsuarioNoEncontradoException{
         if(usuarioRepositorio.existsById(id)){
             usuarioRepositorio.save(usuario);
+        }else{
+            throw new UsuarioNoEncontradoException("ese usuario no existe");
         }
     }
 
@@ -72,7 +84,4 @@ public class UsuarioServicio {
 
         return Optional.of(tarjetas);
     }
-
-
-    
 }
